+++ packages/expo-notifications/src/NotificationPermissions.ts
@@ -1,13 +1,21 @@
-import { Platform } from '@unimodules/core';
+import { Platform, UnavailabilityError } from '@unimodules/core';
 
 import { NotificationPermissionsRequest } from './NotificationPermissions.types';
 import NotificationPermissionsModule from './NotificationPermissionsModule';
 
 export async function getPermissionsAsync() {
+  if (!NotificationPermissionsModule.getPermissionsAsync) {
+    throw new UnavailabilityError('Notifications', 'getPermissionsAsync');
+  }
+
   return await NotificationPermissionsModule.getPermissionsAsync();
 }
 
 export async function requestPermissionsAsync(permissions?: NotificationPermissionsRequest) {
+  if (!NotificationPermissionsModule.requestPermissionsAsync) {
+    throw new UnavailabilityError('Notifications', 'requestPermissionsAsync');
+  }
+
   const requestedPermissions = permissions ?? {
     ios: {
       allowAlert: true,
