+++ packages/expo-notifications/src/NotificationPermissionsModule.ts
@@ -1,15 +1,5 @@
-import { NativeModulesProxy, ProxyNativeModule } from '@unimodules/core';
+import { NativeModulesProxy } from '@unimodules/core';
 
-import {
-  NotificationPermissionsStatus,
-  NativeNotificationPermissionsRequest,
-} from './NotificationPermissions.types';
-
-export interface NotificationPermissionsModule extends ProxyNativeModule {
-  getPermissionsAsync: () => Promise<NotificationPermissionsStatus>;
-  requestPermissionsAsync: (
-    request: NativeNotificationPermissionsRequest
-  ) => Promise<NotificationPermissionsStatus>;
-}
+import { NotificationPermissionsModule } from './NotificationPermissionsModule.types';
 
 export default (NativeModulesProxy.ExpoNotificationPermissionsModule as any) as NotificationPermissionsModule;
