+++ packages/expo-notifications/src/NotificationPermissionsModule.web.ts
@@ -1,0 +1,6 @@
+import { NotificationPermissionsModule } from './NotificationPermissionsModule.types';
+
+export default {
+  addListener: () => {},
+  removeListeners: () => {},
+} as NotificationPermissionsModule;
